+++ tests/src/io/segment/android/test/TestCases.java
@@ -5,8 +5,9 @@ import io.segment.android.models.BasePayload;
 import io.segment.android.models.Batch;
 import io.segment.android.models.Context;
 import io.segment.android.models.EasyJSONObject;
-import io.segment.android.models.EventProperties;
+import io.segment.android.models.Group;
 import io.segment.android.models.Identify;
+import io.segment.android.models.Props;
 import io.segment.android.models.Providers;
 import io.segment.android.models.Screen;
 import io.segment.android.models.Track;
@@ -17,6 +18,7 @@ import java.util.GregorianCalendar;
 import java.util.LinkedList;
 import java.util.Random;
 import java.util.TimeZone;
+import java.util.UUID;
 
 @SuppressWarnings("serial")
 public class TestCases {
@@ -24,6 +26,7 @@ public class TestCases {
 	public static Calendar calendar;
 	
 	public static Identify identify;
+	public static Group group;
 	public static Track track;
 	public static Screen screen;
 	public static Alias alias;
@@ -40,7 +43,9 @@ public class TestCases {
 		calendar.set(Calendar.MINUTE, 20);
 		calendar.set(Calendar.SECOND, 15);
 		
-		identify = new Identify("ilya@segment.io", 
+		identify = new Identify(
+				UUID.randomUUID().toString(),
+				"ilya@segment.io", 
 				new Traits(
 						"name","Achilles", 
 						"email", "achilles@segment.io", 
@@ -50,9 +55,21 @@ public class TestCases {
 							.put("name", "Company, inc.")), 
 				calendar,
 				new Context().setIp("192.168.1.1"));
+
+		group = new Group(
+				UUID.randomUUID().toString(),
+				"ilya@segment.io", 
+				"segmentio_id",
+				new Traits(
+					"name","Segment.io", 
+					"plan", "Premium"), 
+				calendar,
+				new Context().setIp("192.168.1.1"));
 		
-		track = new Track("ilya@segment.io", "Played a Song on Android", 
-				new EventProperties(
+		track = new Track(
+				UUID.randomUUID().toString(),
+				"ilya@segment.io", "Played a Song on Android", 
+				new Props(
 						"name", "Achilles",
 						"revenue", 39.95,
 						"shippingMethod", "2-day"),
@@ -65,8 +82,11 @@ public class TestCases {
 								.setEnabled("KISSMetrics", true)
 								.setEnabled("Google Analytics", true)));
 		
-		screen = new Screen("ilya@segment.io", "Login Page", 
-				new EventProperties(
+		screen = new Screen(
+				UUID.randomUUID().toString(),
+				"ilya@segment.io", 
+				"Login Page", 
+				new Props(
 						"logged-in", true,
 						"type", "teacher"),
 						calendar, 
@@ -94,11 +114,15 @@ public class TestCases {
 	
 	public static BasePayload random() {
 		
-		switch(random.nextInt(3)) {
+		switch(random.nextInt(5)) {
 			case 0:
 				return TestCases.identify;
 			case 1: 
 				return TestCases.track;
+			case 2: 
+				return TestCases.group;
+			case 3: 
+				return TestCases.screen;
 			default:
 				return TestCases.alias;
 		}